+++ components/bottombar/impl/src/main/java/com/flipperdevices/bottombar/impl/composable/ComposableInAppNotification.kt
@@ -1,35 +1,25 @@
 package com.flipperdevices.bottombar.impl.composable
 
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.collectAsState
-import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
 import com.flipperdevices.bottombar.impl.viewmodel.InAppNotificationState
-import com.flipperdevices.bottombar.impl.viewmodel.InAppNotificationViewModel
-import com.flipperdevices.core.ktx.android.OnLifecycleEvent
 import com.flipperdevices.inappnotification.api.InAppNotificationRenderer
-import tangle.viewmodel.compose.tangleViewModel
+import com.flipperdevices.inappnotification.api.model.InAppNotification
 
 @Composable
 fun ComposableInAppNotification(
     notificationRenderer: InAppNotificationRenderer,
-    modifier: Modifier = Modifier,
-    notificationViewModel: InAppNotificationViewModel = tangleViewModel(),
+    notificationState: InAppNotificationState,
+    onNotificationHidden: (notification: InAppNotification) -> Unit,
+    modifier: Modifier = Modifier
 ) {
-    val notificationState by notificationViewModel.state().collectAsState()
-
-    OnLifecycleEvent(onEvent = notificationViewModel::onLifecycleEvent)
-
-    val localNotificationState = notificationState
-    if (localNotificationState !is InAppNotificationState.ShownNotification) {
+    if (notificationState !is InAppNotificationState.ShownNotification) {
         return
     }
     notificationRenderer.InAppNotification(
-        notification = localNotificationState.notification,
+        notification = notificationState.notification,
         modifier = modifier
     ) {
-        notificationViewModel.onNotificationHidden(
-            notification = localNotificationState.notification
-        )
+        onNotificationHidden(notificationState.notification)
     }
 }