+++ components/bottombar/impl/src/main/java/com/flipperdevices/bottombar/impl/composable/bottombar/BottomTabMapper.kt
@@ -3,27 +3,27 @@ package com.flipperdevices.bottombar.impl.composable.bottombar
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.res.stringResource
 import com.flipperdevices.bottombar.impl.R
-import com.flipperdevices.bottombar.impl.model.FlipperBottomTab
+import com.flipperdevices.bottombar.impl.model.BottomBarNavigationConfig
 import com.flipperdevices.bottombar.model.TabState
-import com.flipperdevices.connection.api.ConnectionApi
 import com.flipperdevices.core.ui.theme.LocalPallet
 
 @Composable
 fun getTabStateFromFlipperBottomTab(
-    connectionApi: ConnectionApi,
-    bottomTab: FlipperBottomTab,
+    connectionTabState: TabState,
+    bottomTab: BottomBarNavigationConfig,
     hubHasNotification: Boolean
 ): TabState {
     return when (bottomTab) {
-        FlipperBottomTab.DEVICE -> connectionApi.getConnectionTabState()
-        FlipperBottomTab.ARCHIVE -> TabState.Static(
+        BottomBarNavigationConfig.Device -> connectionTabState
+        BottomBarNavigationConfig.Archive -> TabState.Static(
             selectedIcon = R.drawable.ic_archive_selected,
             notSelectedIcon = R.drawable.ic_archive_unselected,
             text = stringResource(R.string.bar_title_archive),
             selectedColor = LocalPallet.current.bottomBarSelected,
             unselectedColor = LocalPallet.current.bottomBarUnselected
         )
-        FlipperBottomTab.HUB -> TabState.Static(
+
+        BottomBarNavigationConfig.Hub -> TabState.Static(
             selectedIcon = R.drawable.ic_hub_filled,
             notSelectedIcon = R.drawable.ic_hub,
             text = stringResource(R.string.bar_title_hub),