+++ components/bottombar/impl/src/main/java/com/flipperdevices/bottombar/impl/viewmodel/BottomBarViewModel.kt
@@ -9,9 +9,9 @@ import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.onEach
 import kotlinx.coroutines.plus
-import tangle.viewmodel.VMInject
+import javax.inject.Inject
 
-class BottomBarViewModel @VMInject constructor(
+class BottomBarViewModel @Inject constructor(
     hubApi: HubApi
 ) : ViewModel() {
     private val hasNotificationHubStateFlow = MutableStateFlow(false)