+++ components/info/api/src/main/java/com/flipperdevices/info/api/screen/InfoFeatureEntry.kt
@@ -1,9 +1,0 @@
-package com.flipperdevices.info.api.screen
-
-import com.flipperdevices.core.ui.navigation.AggregateFeatureEntry
-import com.flipperdevices.core.ui.navigation.FeatureScreenRootRoute
-
-interface InfoFeatureEntry : AggregateFeatureEntry {
-    override val ROUTE: FeatureScreenRootRoute
-        get() = FeatureScreenRootRoute.DEVICE_INFO
-}